Return a consumer-id that can be used to stop it
Clients should no longer assume there is a future being used, and in fact we no longer return it to them. Cancelling it no longer had the effect it once did anyway.

Now you save the consumer-id it returns and then use that as the arg to `stop-consumer` when you want to shut down.