Adds a :no-retries option
If `:no-retries true` is specified, ignores
visibility timeout and just deletes the message
before it processes it. Yeah, this is dangerous,
but in our practice we sometimes/often have 
messages that take longer than the maximum
visibility timeout of SQS (12 hours), so they
always spawn duplicates and these are the worst
offenders because they have the most data and thus
clog our database up if it happens, say, late on
a Friday.